12.2

+++ src/eval.py
@@ -10,4 +10,4 @@ def f1_score_tensors(y_true, y_pred):
 
 # predict
 def predict_labels(output, threshold=0.25):
-    return (output > threshold).type(torch.uint8)
+    return (output > threshold)
